+++ src/js/validator/phone.js
@@ -107,8 +107,7 @@
                     break;
                     
         		case 'RO':
-        		    // Phone intl/national
-        		    // All mobile network and land line.
+        		    // All mobile network and land line
         		    isValid = (/^(\+4|)?(07[0-8]{1}[0-9]{1}|02[0-9]{2}|03[0-9]{2}){1}?(\s|\.|\-)?([0-9]{3}(\s|\.|\-|)){2}$/g).test(value);
                     break;